refactor(scripts): combine score analysis scripts into one
Merges ingest_score_analysis.py and ingest_bm25_param_search.py into
ingest_search_score_analysis.py with two subcommands:
  analyze      — BM25 + semantic similarity distributions and coverage curve
  paramsearch  — k1 / title_boost grid search heatmaps
